requirements.txt
django>=6.0
djangorestframework>=3.17
djangorestframework-simplejwt>=5.5
django-cors-headers>=4.9
celery>=5.6
redis>=7.4
psycopg2-binary>=2.9
gunicorn>=23.0
python-dotenv>=1.2
django-rosetta>=0.9
Pillow>=10.0
requests>=2.31
django-celery-beat>=2.6
whitenoise>=6.0
